+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Yii Framework 2 debug extension Change Log
 - Enh #16: Added ability to EXPLAIN queries in Database panel for MySQL, SQLite, PostgreSQL and Cubrid (laszlovl, samdark)
 - Enh #19: Mark selected log item in dropdown list with bold font and an arrow (idMolotov)
 - Enh #25: Make use of full screen width in debug toolbar backend (dynasource, cebe)
+- Enh #36: Added check for EXPLAIN support in DbPanel (webdevsega)
 - Enh: More compact toolbar (samdark)
 - Enh: Display colorful status at index page (samdark)
 - Enh: More readable format for date and time at index page (samdark)